requirement.txt
fastapi
uvicorn
pillow
httpx
aiofiles
python-multipart